Improve `canTunnel` Part of microsoft/vscode-internalbacklog#1709
