+++ src/vs/workbench/services/remote/electron-browser/tunnelServiceImpl.ts
@@ -9,14 +9,21 @@ import { ISignService } from 'vs/platform/sign/common/sign';
 import { IRemoteAgentService } from 'vs/workbench/services/remote/common/remoteAgentService';
 import { BaseTunnelService } from 'vs/platform/remote/node/tunnelService';
 import { nodeSocketFactory } from 'vs/platform/remote/node/nodeSocketFactory';
+import { IWorkbenchEnvironmentService } from 'vs/workbench/services/environment/common/environmentService';
+import { URI } from 'vs/base/common/uri';
 
 export class TunnelService extends BaseTunnelService {
 	public constructor(
 		@ILogService logService: ILogService,
 		@ISignService signService: ISignService,
 		@IProductService productService: IProductService,
-		@IRemoteAgentService _remoteAgentService: IRemoteAgentService
+		@IRemoteAgentService _remoteAgentService: IRemoteAgentService,
+		@IWorkbenchEnvironmentService private environmentService: IWorkbenchEnvironmentService
 	) {
 		super(nodeSocketFactory, logService, signService, productService);
+	}
+
+	canTunnel(uri: URI): boolean {
+		return super.canTunnel(uri) && !!this.environmentService.remoteAuthority;
 	}
 }
